Appending a version query parameter at end of potree.js script file causes the script to fail to load properly
To avoid the browser caching JavaScript source files, it is common to append a version or hash parameter (Example: <script src="~/potree/potree/potree.js?version=myversion123">).  Previous to this commit, this causes the module loading to fail.  This commit strips the query parameters so that the loading will work properly.